requirements.txt
openai-whisper
ffmpeg-python
tqdm